move web service handlers into main file for importing
